fix(daemon): Make directory `move()` compatible with Exo name hubs
Adds a more complete test suite for the directory's `move()`. In the
process of adding these tests, it was discovered that the method was
incompatible with Exo name hubs, because we did not call `E()` on the
`fromHub` or `toHub` values. This has been amended, and test cases have
been added.