feat(compartment-mapper): add a prop for custom fields in policy
This adds an optional property, `options`, to the `PackagePolicy` type.

Here, consumers can add any fields or metadata they need to the policy. Any values will pass through the policy validator _alive, and unspoiled_.

Endo should not read the contents of this property directly.  The default type of `unknown` means that access to the value must have an explicit type assertion; it is the responsibility of the consumer to provide a proper type for this field.