feat(import-bundle): accept symbol-named properties on inescapableGlobalProperties
Previously, the `inescapableGlobalProperties` option to
`importBundle()` only accepted enumerable string-named
properties. This commit changes that to accept all properties (even
unenumerable ones), and to accept symbol-named properties (instead of
only string-named ones).

We'll use this to provide `passStyleOf` to the vat environment built
by liveslots, so it can use a real WeakMap, instead of the (slower)
"virtual object -aware WeakMap".

closes #2376