fix(ses): make test insensitive to an irrelevant env var (#2403)
Closes: #XXXX
Refs: #2383 

## Description

Of the various property override tests, the ones testing the default
`'moderate'` setting were calling `lockdown` without an `overrideTaming`
option, counting on it to default to `'moderate'`. However, if run
locally in an environment where the environment variable
`LOCKDOWN_OVERRIDE_TAMING` was set to something else, the test would
pointlessly misbehave.

This PR merely sets the option explicitly for those cases, so the tests
for the `'moderate'` case are as insensitive to environment variables as
the tests for the other cases.

### Security Considerations

none
### Scaling Considerations

none
### Documentation Considerations

none
### Testing Considerations

The point. The code before this PR would always work correctly under CI.
But the test may pointless fail locally depending on the developers
environment variable settings.
### Compatibility Considerations

none
### Upgrade Considerations

none