fix(compartment-mapper): fix #2407 include symbol-named properties (#2408)
Closes: #2407 
Refs: https://github.com/Agoric/agoric-sdk/pull/9874
https://github.com/Agoric/agoric-sdk/pull/9431
https://github.com/Agoric/agoric-sdk/pull/9781 #2377

## Description

While working on https://github.com/Agoric/agoric-sdk/pull/9874 , I
found that it still failed to endow `passStyleOf` via a symbol-named
property as recently agree.

Turns out there was still a lurking `Object.keys` on the compartment
endowment pathway that needs to be changed to enumerate all enumerable
own properties, whether string-named or symbol-named.

### Security Considerations

None

### Scaling Considerations

For this PR, none. But it enables liveslots to endow its `passStyleOf`
to the compartments it makes, which has the scaling benefits explained
at https://github.com/Agoric/agoric-sdk/issues/9781

### Documentation Considerations

For the normal developer, none. Once endowed, `passStyleOf` will have
performance that is less surprising.

### Testing Considerations

https://github.com/Agoric/agoric-sdk/pull/9874 both patches in the
equivalent of this PR, uses it to endow `passStyleOf`, and test that it
has done so, across both node and XS.

However, after this fix, https://github.com/Agoric/agoric-sdk/pull/9874
works on the Node host. But it still fails on the XS host for
undiagnosed reasons, so there may be another bug like #2407 still
lurking somewhere. Until diagnosed, it isn't clear if that remaining
problem is an endo problem or an agoric-sdk problem.

### Compatibility and Upgrade Considerations

Except for code meant only to test this, `passStyleOf` should not have
any observable difference. We have not attempted to make an other use of
symbol-named global properties, so there should be no compat or upgrade
issues.