fix(compartment-mapper): disallow dynamic requires of arbitrary, unknown paths
This changes the logic so that this particular class of dynamic requires is disallowed.

If the specifier is not accessible per policy, the usual policy-enforcement error message applies.  If the specifier is not accessible because it's nowhere in the exit module, the error will mention that it can't load an unknown module.

Also fixes:

- dynamic requires of exit modules (e.g., Node.js builtins)
- dynamic requires made from the entry compartment descriptor
- some peer review comments