refactor(pass-style): Avoid name convention conflict (#2666)
closes: #XXXX
https://github.com/Agoric/agoric-sdk/pull/10751

## Description

We generally use names of the form `^([A-Z]\w*)I$`, i.e., identifiers
beginning with a capital letter and ending with a lone capital "I", for
interface guards. However there were some uses of the same form of name
for typescript interfaces. Because interfaces and interface guards are
adjacent concepts, this could be confusing. This PR fixes this for endo.
The companion https://github.com/Agoric/agoric-sdk/pull/10751 fixes it
for agoric-sdk. Because the endo names in question were not exported,
there is clearly no dependency between these two PRs.

### Security Considerations

Consistent naming conventions makes code more reviewable, which is good
for security.

### Scaling Considerations
none

### Documentation Considerations
Consistent naming conventions helps documentation and its readers.


### Testing Considerations
none

### Compatibility Considerations

Because these names are not exported, and are only type names erased
prior to execution, none.

### Upgrade Considerations
Because these are only type names that are erased prior to execution,
none.