fix(compartment-mapper)!: invert keys/values of renames in captureFromMap
BREAKING CHANGE: The `CaptureResult.compartmentRenames` record returned by `captureFromMap()` is now _new to old_; previously it was _old to new_.

This more accurately solves for the use case of `compartmentRenames`, because the main use case demands the record is inverted. We can plainly see this issue in `captureFromMap`'s test suite (which this change simplifies).

In a future change, [branded types](https://www.learningtypescript.com/articles/branded-types) could describe the distinction between key and value, which are currently both `string`.

> [!IMPORTANT]
>
> The only consumer of `captureFromMap` that I'm aware of is _me_, and _me_ doesn't need this to be SemVer major.